requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
langchain==0.3.1
langchain-openai==0.2.1
langchain-community==0.3.1
faiss-cpu==1.8.0
pymupdf==1.24.10
tavily-python==0.5.0
python-dotenv==1.0.1
